Probleme in den Randbereichen beseitigt

+++ src/Resources/contao/classes/pdfNcAttachmentHooks.php
@@ -13,30 +13,47 @@ namespace Softleister\Pdfncattachment;
 require_once( TL_ROOT . '/vendor/do-while/contao-pdf-nc-attachment-bundle/src/Resources/contao/classes/pdfnc_helper.php' );
 
 //-----------------------------------------------------------------
-//  pdfNcAttachmentHooks:    Ausgabe des PDF
+//  pdfNcAttachmentHooks:    Output of the PDF
 //-----------------------------------------------------------------
 class pdfNcAttachmentHooks extends \Backend
 {
     protected   $arrProtectflags = array('modify','extract','print','print-high','copy','annot-forms','fill-forms');
 
     //-----------------------------------------------------------------
-    //  execute:    Maildaten verarbeiten
+    //  execute:    Process mail data
     //-----------------------------------------------------------------
     public function execute($objMessage, &$arrTokens, $language, $objGatewayModel)
     {
-        if( $objMessage->gateway_type !== 'email' ) return true;        // nur für den Gateway-Typ "email"
-        if( $objGatewayModel->pdfnc_on != 1 ) return true;              // "PDF-Formular ausfüllen" inaktiv
+        if( $objMessage->gateway_type !== 'email' ) return true;        // for gateway type "email" only
+        if( $objGatewayModel->pdfnc_on != 1 ) return true;              // "Fill in PDF form" inactive
 
         $filename = standardize(\StringUtil::restoreBasicEntities($objMessage->title)) . \Haste\Util\StringUtil::recursiveReplaceTokensAndTags( $objGatewayModel->pdfnc_fileext, $arrTokens );
         
+        //--- member directory? ---
         $savepath = \FilesModel::findByUuid($objGatewayModel->pdfnc_savepath)->path;
-        if( file_exists(TL_ROOT . '/' . $savepath . '/' . $filename . '.pdf') ) {
-            $i = 2;
-            while( file_exists(TL_ROOT . '/' . $savepath . '/' . $filename . '-' . $i . '.pdf') ) $i++;
-            $filename = $filename . '-' . $i;
+        if( $objGatewayModel->pdfnc_useHomeDir && FE_USER_LOGGED_IN ) {
+			$this->import('FrontendUser', 'User');
+
+			if( $this->User->assignDir && $this->User->homeDir ) {
+                $dir = \FilesModel::findByUuid($this->User->homeDir)->path;
+                if( is_dir( TL_ROOT . '/' . $dir ) ) {
+				    $savepath = $dir;                                   // Accept member directory
+                }
+			}
         }
-        $ttt = TL_ROOT . '/' . $savepath . '/' . $filename . '.pdf';
-        file_put_contents(TL_ROOT . '/' . $savepath . '/' . $filename . '.pdf', '');  // leere Datei erzeugen
+
+        if( file_exists(TL_ROOT . '/' . $savepath . '/' . $filename . '.pdf') ) {
+            if( !$objGatewayModel->pdfnc_overwrite ) {
+                $i = 2;
+                while( file_exists(TL_ROOT . '/' . $savepath . '/' . $filename . '-' . $i . '.pdf') ) $i++;
+                $filename = $filename . '-' . $i;
+
+                file_put_contents(TL_ROOT . '/' . $savepath . '/' . $filename . '.pdf', '');    // create empty file
+            }
+        }
+        else {
+            file_put_contents(TL_ROOT . '/' . $savepath . '/' . $filename . '.pdf', '');        // create empty file
+        }
 
         $arrPDF = array( 'gateid'        => $objGatewayModel->id,
                          'gatetitle'     => $objGatewayModel->title,
@@ -54,11 +71,11 @@ class pdfNcAttachmentHooks extends \Backend
                          'title'         => $objGatewayModel->pdfnc_title,
                          'author'        => $objGatewayModel->pdfnc_author,
                          'tokenlist'     => $objGatewayModel->pdfnc_tokens,
+                         'arrTokens'     => $arrTokens
                         );
-        unset( $arrFields );
         if( !is_array($arrPDF['protectflags']) ) $arrPDF['protectflags'] = array( $arrPDF['protectflags'] );
 
-        // Offsets eintragen, wenn angegeben
+        // Enter offsets if specified
         $ofs = deserialize($objGatewayModel->pdfnc_offset);
         if( isset($ofs[0]) && is_numeric($ofs[0]) ) $arrPDF['offset'][0] = $ofs[0];
         if( isset($ofs[1]) && is_numeric($ofs[1]) ) $arrPDF['offset'][1] = $ofs[1];
@@ -67,28 +84,54 @@ class pdfNcAttachmentHooks extends \Backend
         if( isset($GLOBALS['TL_HOOKS']['pdfnc_BeforePdf']) && \is_array($GLOBALS['TL_HOOKS']['pdfnc_BeforePdf']) ) {
             foreach( $GLOBALS['TL_HOOKS']['pdfnc_BeforePdf'] as $callback ) {
                 $arrPDF = \System::importStatic($callback[0])->{$callback[1]}( $arrPDF, $this );
+                $arrTokens = $arrPDF['arrTokens'];
             }
         }
 
 
-        //--- PDF-Datei erstellen und speichern ---
+        //-- Include Settings
+        $tcpdfinit = \Config::get("pdftemplateTcpdf");
+
+        // 1: Own settings addressed via app/config/config.yml
+        if( !empty($tcpdfinit) && file_exists(TL_ROOT . '/' . $tcpdfinit) ) {
+            require_once(TL_ROOT . '/' . $tcpdfinit);
+        }
+        // 2: Own tcpdf.php from files directory
+        else if( file_exists(TL_ROOT . '/files/tcpdf.php') ) {
+            require_once(TL_ROOT . '/files/tcpdf.php');
+        }
+        // 3: From config directory (up to Contao 4.6)
+        else if( file_exists(TL_ROOT . '/vendor/contao/core-bundle/src/Resources/contao/config/tcpdf.php') ) {
+            require_once(TL_ROOT . '/vendor/contao/core-bundle/src/Resources/contao/config/tcpdf.php');
+        }
+        // 4: From config directory of tcpdf-bundle (from Contao 4.7)
+        else if( file_exists(TL_ROOT . '/vendor/contao/tcpdf-bundle/src/Resources/contao/config/tcpdf.php') ) {
+            require_once(TL_ROOT . '/vendor/contao/tcpdf-bundle/src/Resources/contao/config/tcpdf.php');
+        }
+        // 5: not found? - Then take it from this extension
+        else {
+            require_once(TL_ROOT . '/vendor/do-while/contao-pdf-nc-attachment-bundle/src/Resources/contao/config/tcpdf.php');
+        }
+
+
+        //--- Create and save PDF file ---
         $arrTokens['pdfnc_attachment'] = $arrTokens['pdfnc_document'] = '';
         $pdfdatei = $savepath . '/' . $filename . '.pdf';
 
         if( pdfnc_helper::getPdfData( $arrPDF, $arrTokens, $pdfdatei ) ) {
 
-            //--- PDF-Datei in der Dateiverwaltung eintragen ---
-            $objFile = \Dbafs::addResource( $pdfdatei );                    // Datei in der Dateiverwaltung eintragen
+            //--- Enter PDF file in the file manager ---
+            $objFile = \Dbafs::addResource( $pdfdatei );
             \Dbafs::updateFolderHashes( $strUploadFolder );
 
-            //--- Token für erstellte Datei mitgeben ---
+            //--- Create token for created file ---
             $arrTokens['pdfnc_attachment'] = $pdfdatei;
             $arrTokens['pdfnc_document'] = basename( $pdfdatei );
 
-            //--- Eintrag im Log ---
+            //--- Entry in log ---
             \System::Log('PDF attachment "' . $filename . '.pdf" has been created', __METHOD__, TL_ACCESS);
         }
-        else $pdfdatei = '';        // es wurde keine Datei erzeugt
+        else $pdfdatei = '';        // no file was created
 
 
         // HOOK: after pdf generation
@@ -97,7 +140,7 @@ class pdfNcAttachmentHooks extends \Backend
                 \System::importStatic($callback[0])->{$callback[1]}( $pdfdatei, $arrPDF, $this );
             }
         }
-        return true;                // Benachrichtigung darf gesendet werden
+        return !isset( $arrTokens['do_not_send_notification'] );                // Notification may be sent
     }
